requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
WTForms==3.0.1
pandas==2.0.3
reportlab==4.0.4
email_validator==2.1.0
requests==2.31.0